image_gen/requirements.txt
astropy==5.3.4
cupy_cuda11x==13.2.0
matplotlib==3.8.0
numpy==2.0.1
scikit_learn==1.2.2
scipy==1.14.0
tqdm==4.65.0
